+++ Engine/Source/Editor/BlueprintGraph/Private/K2Node_EaseFunction.cpp
@@ -481,7 +481,7 @@ void UK2Node_EaseFunction::RefreshPinVisibility()
 	const auto EaseFuncPin = GetEaseFuncPin();
 	UEnum * Enum = StaticEnum<EEasingFunc::Type>();
 	check(Enum != NULL);
-	const int32 NewEasingFunc = CanCustomizeCurve() ? Enum->GetValueByName(*EaseFuncPin->DefaultValue) : -1;
+	const int64 NewEasingFunc = CanCustomizeCurve() ? Enum->GetValueByName(*EaseFuncPin->DefaultValue) : -1;
 	
 	// Early exit in case no changes are required
 	const UEdGraphSchema_K2* K2Schema = Cast<const UEdGraphSchema_K2>(GetSchema());
