+++ Engine/Source/Editor/BlueprintGraph/Public/BlueprintNodeTemplateCache.h
@@ -75,7 +75,7 @@ public:
 	 * 
 	 * @return The approximated total (in bytes) that this cache has allocated.
 	 */
-	int32 GetEstimateCacheSize() const;
+	int64 GetEstimateCacheSize() const;
 
 	/**
 	 * External systems can make changes that alter the memory footprint of the
@@ -85,7 +85,7 @@ public:
 	 * 
 	 * @return The new approximated total (in bytes) that this cache has allocated.
 	 */
-	int32 RecalculateCacheSize();
+	int64 RecalculateCacheSize();
 
 	// FGCObject interface
 	virtual void AddReferencedObjects(FReferenceCollector& Collector) override;
@@ -136,5 +136,5 @@ private:
 	 * entry is added, so we keep this approximate tally of memory allocated for
 	 * UObjects (owned by this system).
 	 */
-	int32 ApproximateObjectMem;
+	int64 ApproximateObjectMem;
 };
